Impossible power selection now gives an error
For example, giving red power to a ship that cannot capture anything
gives an error and cancels the selection.
Previously, the game would wait for a capture selection that could not be given
and the player would have to click cancel.
Error message and cancellation also occur for blue and yellow power
given to ships that cannot trade and move respectively.